LayoutTests/imported/w3c/web-platform-tests/navigation-timing/test_navigation_type_enums-expected.txt

Description

This test validates that the TYPE_* enumerations of window.performance.navigation exist and their values are correct.


PASS window.performance is defined
PASS window.performance.navigation is defined
PASS window.performance.navigation.TYPE_NAVIGATE is defined.
PASS window.performance.navigation.TYPE_NAVIGATE = 0
PASS window.performance.navigation.TYPE_RELOAD is defined.
PASS window.performance.navigation.TYPE_RELOAD = 1
PASS window.performance.navigation.TYPE_BACK_FORWARD is defined.
PASS window.performance.navigation.TYPE_BACK_FORWARD = 2
PASS window.performance.navigation.TYPE_RESERVED is defined.
PASS window.performance.navigation.TYPE_RESERVED = 255
